Revert "fix(tasks): keep processing available tasks"
This reverts commit 6c02c74425241313994fa4663e9edf7f5a044ea3.

+++ komga/src/main/kotlin/org/gotson/komga/application/tasks/TaskProcessor.kt
@@ -65,18 +65,16 @@ class TaskProcessor(
   }
 
   private fun takeAndProcess() {
-    while (true) {
-      logger.debug { "Try to process first available task" }
-      val task =
-        tasksRepository.takeFirst() ?: run {
-          logger.debug { "No available task found" }
-          return
-        }
-
+    logger.debug { "Try to process first available task" }
+    val task = tasksRepository.takeFirst()
+    if (task != null) {
       logger.debug { "Found task to process: $task" }
       taskHandler.handleTask(task)
       logger.debug { "Task processed, remove it from the queue: $task" }
       tasksRepository.delete(task.uniqueId)
+      processAvailableTask()
+    } else {
+      logger.debug { "No available task found" }
     }
   }
 }
